refactor(ledger store): Separate best, staged, & staking ledger stores
