+++ rust/src/command/store.rs
@@ -77,22 +77,22 @@ pub trait UserCommandStore {
     ///////////////
 
     /// Iterator for user commands via global slot
-    fn user_commands_slot_iterator<'a>(&'a self, mode: IteratorMode) -> DBIterator<'a>;
+    fn user_commands_slot_iterator(&self, mode: IteratorMode) -> DBIterator<'_>;
 
     /// Iterator for user commands via blockchain length
-    fn user_commands_height_iterator<'a>(&'a self, mode: IteratorMode) -> DBIterator<'a>;
+    fn user_commands_height_iterator(&self, mode: IteratorMode) -> DBIterator<'_>;
 
     /// Iterator for user commands by sender via block height
-    fn txn_from_height_iterator<'a>(&'a self, mode: IteratorMode) -> DBIterator<'a>;
+    fn txn_from_height_iterator(&self, mode: IteratorMode) -> DBIterator<'_>;
 
     /// Iterator for user commands by sender via global slot
-    fn txn_from_slot_iterator<'a>(&'a self, mode: IteratorMode) -> DBIterator<'a>;
+    fn txn_from_slot_iterator(&self, mode: IteratorMode) -> DBIterator<'_>;
 
     /// Iterator for user commands by sender via block height
-    fn txn_to_height_iterator<'a>(&'a self, mode: IteratorMode) -> DBIterator<'a>;
+    fn txn_to_height_iterator(&self, mode: IteratorMode) -> DBIterator<'_>;
 
     /// Iterator for user commands by receiver via global slot
-    fn txn_to_slot_iterator<'a>(&'a self, mode: IteratorMode) -> DBIterator<'a>;
+    fn txn_to_slot_iterator(&self, mode: IteratorMode) -> DBIterator<'_>;
 
     /////////////////////////
     // User command counts //
